Review suggestions for #5370, running on windows (#5374)
