+++ etc/testing/test_worker.sh
@@ -1,12 +1,14 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set -euxo pipefail
 
-export TIMEOUT="${1}"
-PROM_PORT="$(
-  kubectl --namespace=monitoring get svc/prometheus -o json \
-    | jq -r .spec.ports[0].nodePort
-)"
-export PROM_PORT
+TIMEOUT="120s"
+if [ "$#" -eq 1 ]; then
+  TIMEOUT="${1}"
+elif [ "$#" -gt 1 ]; then
+  echo "Too many parameters, expected 1, got $#"
+  echo "Usage: $0 <TIMEOUT>"
+  exit 1
+fi
 
 # grep out PFS server logs, as otherwise the test output is too verbose to
 # follow and breaks travis
